3/4

+++ website.com/admin/product/delete.php
@@ -1,21 +1,22 @@
 <?php
     session_start();
-    if(!isset($_SESSION["is_login"]) || $_SESSION["is_login"] !== true){
-        echo 'Không đc phép vô đây';die();
-    }
     //Kết nối SQL
     require_once ('./../libs/database.php');
     $connect = connect_db();
     
-    $id = $_GET['id'];
-    $sql = "DELETE FROM product WHERE id = " . $id;
-    $result = mysqli_query($connect, "SELECT * FROM product WHERE id =  $id");
-    $row   = mysqli_fetch_assoc($result);
-    $old_image = './../uploads/' .$row['image'];
-    unlink($old_image);
-    mysqli_query($connect, $sql);
-    session_start();
-    $_SESSION["message"] = "Đã xóa thành công id = $id";
-    header('Location: http://localhost/training-php/website.com/product');
+    $id = isset($_GET['id']) ? $_GET['id'] : 0;
+
+    if($id > 0) {
+        $sql = "DELETE FROM product WHERE id = " . $id;
+        $result = mysqli_query($connect, "SELECT * FROM product WHERE id =  $id");
+        $row   = mysqli_fetch_assoc($result);
+        $old_image = './../uploads/' .$row['image'];
+        unlink($old_image);
+        mysqli_query($connect, $sql);
+        $_SESSION["message"] = "Đã xóa thành công id = $id";
+    }
+
+    header("Location: $base_url/admin/index.php?controller=product&action=index");
+
 ?>
